if enable_smartfilter is set to no, there is no filter_repeatlimit in the prefs generated page, which causes saveprefs to fail with undefined $value
git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@651 9f51437f-89f6-0310-a5ff-ff9a3806d2ee

+++ src/cgi-bin/openwebmail/openwebmail-prefs.pl
@@ -1888,7 +1888,7 @@ sub saveprefs {
                last;
             }
          }
-      } elsif ($key eq 'filter_repeatlimit') {
+      } elsif ($config{enable_smartfilter} && ($key eq 'filter_repeatlimit')) {
          # if repeatlimit changed, redo filtering may be needed
          unlink(dotpath('filter.check')) if $value != $prefs{filter_repeatlimit};
          $newprefs{$key} = $value;
